Only start loading of deferred assets after UI is ready
