+++ public/index.html
@@ -35,7 +35,7 @@
     <meta property="og:image" content="images/logo_transparent_512x512.png">
     <!-- Resources -->
     <link rel="preload" href="lang/en.json" as="fetch">
-    <link rel="stylesheet" type="text/css" href="styles/main-styles.css">
+    <link rel="stylesheet" type="text/css" href="styles/styles-main.css">
     <link rel="manifest" href="manifest.json">
 </head>
 
@@ -599,13 +599,8 @@
     <script src="scripts/util-main.js"></script>
     <script src="scripts/localization.js"></script>
     <script src="scripts/persistent-storage.js"></script>
+    <script src="scripts/ui-main.js"></script>
     <script src="scripts/main.js"></script>
-    <script defer src="scripts/util.js"></script>
-    <script defer src="scripts/network.js"></script>
-    <script defer src="scripts/ui.js"></script>
-    <script defer src="scripts/qr-code.min.js"></script>
-    <script defer src="scripts/zip.min.js"></script>
-    <script defer src="scripts/no-sleep.min.js"></script>
     <!-- Sounds -->
     <audio id="blop" autobuffer="true">
         <source src="sounds/blop.mp3" type="audio/mpeg">
